pyproject.toml
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/#example
[tool.portray]
output_dir = "documentation_website"
modules = ["tpd"]
port = 8001

[project]
name = "trim-pocket-deployment"
# version = "0.1a"
authors = [
  { name="Strangemother", email="tpd@strangemother.com" },
]
description = "tpd."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = {text='MIT'}
dynamic = ['version']
dependencies = [
    "pyyaml"
]
# dynamic
# entry-points
# gui-scripts
# keywords
# license
# maintainers
# optional-dependencies
# scripts
# urls

[tool.setuptools.package-data]
tpd = ["**/*.js", "**/*.css","**/*.html"]

[project.scripts]
tpd = 'tpd.cli:main'

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
"Homepage" = "http://github.com/strangemother/tpd"
"Bug Tracker" = "http://github.com/strangemother/tpd/issues"
